requirements.txt
# Autres
joblib==1.3.1
pydantic==1.10.7
python-dotenv==1.0.0
scikit_learn==1.3.0


# Packages principales
mysql-connector-python==8.0.33
SQLAlchemy==1.4.7
python-dotenv==1.0.0
psycopg2-binary==2.9.6
PyMySQL==1.0.3
pandas==1.4.4

# Pour API
uvicorn==0.22.0
fastapi==0.95.2